Change: Migrate to Expo 51
This is an initial migration to Expo 51, with some breakages remaining. More commits in future will bring the UI further in line with the original before moving to adding features. Note that the migration also includes moving to the `tamagui` default config, and importing colours manually over directly from a custom config.